+++ src/microsoft/opentelemetry/_constants.py
@@ -44,6 +44,18 @@ _SUPPORTED_INSTRUMENTED_LIBRARIES = (
     "agent_framework",
 )
 
+# Libraries disabled by default when A365 is enabled (agent workloads
+# typically don't need web-framework / HTTP-client instrumentation).
+_A365_DISABLED_INSTRUMENTATIONS = (
+    "django",
+    "fastapi",
+    "flask",
+    "psycopg2",
+    "requests",
+    "urllib",
+    "urllib3",
+)
+
 # --- Console Exporter Constants ---
 
 ENABLE_CONSOLE_ARG = "enable_console"